pyproject.toml
[project]
name = "ai-data-query-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "sqlparse>=0.5.5",
    "streamlit>=1.54.0",
    "tabulate>=0.9.0",
    "xlrd>=2.0.2",
]
